[TASK] Add badge for disabled task group
Currently there is no indication for a user
to see if a task group is disabled/hidden.

This has been changed by adding a badge for
disabled/hidden groups and also another badge
for tasks belonging to a disabled group.
The "late" badge and "next excecution" are now
hidden as well.

Resolves: #99849
Releases: main
Change-Id: I22a44e6940fe38639999a78e6ebc7e6f65ffb7d7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77722
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>