[BUGFIX] Edit buttons using current language-scope in page module
The page module is a language-scoped view of pages. The user is always
in a language context. The language context is respected by all actions
within the module, except the edit button in the module bar. Instead of
respecting the context, a second button is added to enter the translated
page properties.

This is a problem, because the language context is ignored and the user
now faces a button with a different behavior, using the same voice and
tone on their labels.

For example the "view" button does not change the label and openes the
current page always in the selected language. The "edit" button also does
not change the label, but opens always the default language record.

This is a heavy inconsitency in the user experience and needs to be
addressed.

We are now adjusting this to have the edit button always visible and
disable the button if the user has no permission to perform the task.
The edit button will now always respect the the current selected
language and the additional button is therefore dropped.

In the language comparison view, the edit and view buttons are
now removed, because those action buttons already exist on the
corresponding language columns.

Resolves: #99919
Releases: main

Change-Id: I905f39bad275b4d396001ae2be1402eb08e7d160
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77822
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>