[TASK] Refine LiveSearch search demand
The search demand implementation used by the LiveSearch is refined to be
more flexible. Instead of a hard-coded list of properties, the
constructor now receives a list of `DemandProperty` instances that
accept everything that is backed by the `DemandPropertyName` enum,
allowing to pass only the required properties.

The `SearchDemand` class is immutable by design, it's API must not
allow modification by 3rd parties. However, there is an internal
`MutableSearchDemand` acting as a helper that allows to add / override
properties. A `SearchDemand` object may be constructed from a request
object, where it's payload must contain properties from the
aforementioned `DemandPropertyName` enum.

Resolves: #99941
Releases: main
Change-Id: I880855943a67237fa765cd3da94333e8aaa3a2a2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77842
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>