[BUGFIX] Make system information message link binding independent
The link binding of the system information menu messages reuses
module menu identifiers but does not embrace them to avoid additional
attributes on links. We reworked the message link binding to make it
independent from existing link binding and avoid further conflicts.

Instead of refreshing the link bindings after each data fetch, we are
moving the registration for the link click handler to the static parent
element. The link binding now keeps intact after fetching new data.

We currently dont have an encapsulated component for the system
information menu. To avoid nested document queries we are
hardening the selectors for now. In this step we are also removing
jQuery from the system information menu, as there is no use for
this here anymore.

Resolves: #99961
Releases: main
Change-Id: Id312ccc43559d6f630877254bdd09d2d1ed08038
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77868
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>