[TASK] Streamline module link handling
This patch scopes the handling of module links within the module menu
and toolbar. We are now differentiating between module menu events and
module link events and avoiding the double usage of data attributes to
have the scope of the data-attributes assigned.

The internal selectors are now using on data-attributes, this provides
various benefits in contrast to legacy `t3js-...` class based selectors.

- Attributes can contain additional values we need for processing
- We avoid conflicts with for example `*:not([class])` CSS selectors,
  this is helpful to identify links that need special treatment for
  example color adjustments within adaptive components like the
  dropdowns.

While the module menu in its current state still handles the module
links, this patch prepares future work to extract the link management.
Managing the module links in the backend should not be the
responsibility of the module menu.

We are removing all "id" attributes attached to links to avoid conflicts
with other existing identifiers in the backend. The processing has
been adjusted to support this change.

The keyboard navigation is now simplified and adjusted when leaving a
subgroup the selection is now correctly at the parent element and not
the previous parent element.

Resolves: #99972
Releases: main
Change-Id: Ieb87942fc73cab7679c7af939b2a1a247907b6d5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77855
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>