[BUGFIX] Fix PHP 8 warning in LinkValidatorController
The variable $row['url_response']['errorParams'] should always contain
a value. Either the actual error message array or an empty array to
avoid a PHP 8 warning.
The corresponding getErrorMessage methods now have a guard clause to
check if the array is empty to avoid further propagated warnings.
In TYPO3 v13 that signature of the getErrorMessage should be changed to
be array.

The array $errorParams is always initialized to an empty array instead
of null. IN TYPO3 v13 the default value of $errorParams should be []
instead of null.

Resolves: #99988
Releases: main, 11.5
Change-Id: I0f7e10f45e0f5ec7e80b79d705b38fda9a3f1410
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77898
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>