[BUGFIX] Use backend specific visibility aspect in PreviewUriBuilder
The PreviewUriBuilder is used in the backend to generate
frontend URLs. Besides passing the exact page id, it's
also possible to pass the default page id and the desired
language id. In such cases, the PageRouter is responsible
to resolve the id of the translated page. This is internally
done, using the PageRepository. However, due to a missing
visibility context does the PageRepository apply the
whole FrontendRestrictionContainer, including start-
and endtime restrictions.

This is now fixed by passing a backend specific visibility
aspect to the PreviewUriBuilder.

Resolves: #100008
Releases: main, 11.5
Change-Id: Ieacca442588612491f0b3f0d80b30fa64f902e7e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77917
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>