[BUGFIX] Avoid timing issue when loading LiveSearch form
There was a timing issue related to the LiveSearch form.
`Modal.advanced()` invokes an AJAX request, while the event handler for
`typo3-modal-shown` expects that the request sent a response.

This might not be the case when either the network is slow or animations
are disabled as the event is always faster dispatched than the request
can send a response.

The generic handling is now moved into an `ajaxCallback`, executed when
fetching the data finished. To be able to still focus the search field
and mark its content, event handlers for 'modal-loaded' and
'typo3-modal-shown' events are installed, as both events may be
dispatched in any order.

Resolves: #100025
Releases: main
Change-Id: I8aca8ba330af55ce915a414dc5a52ba6eb2af1f1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77941
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>