[BUGFIX] Allow POST requests in Workspace previews
With TYPO3 v9 and site handling, and speaking URLs
that are supported in workspaces, there is no technical
reason to avoid POST requests in Workspaces.

This limitation was added by Kasper in TYPO3 v4.0
in the infamous "tslib_fe" class, whereas the current
code base is much more structured and we can
now safely say, that we have Cookie handling
and POST requests under control via Site Handling,
and middlewares.

This way, forms can be tested and submitted in Workspace
previews.

Resolves: #100051
Releases: main, 11.5
Change-Id: I5ed6afd2384ef685453034f852373bb6e073c974
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77961
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>