[FEATURE] Introduce non-magic repository find methods
This change introduces alternative methods to magic method
__call() of extbase's Repository class. The magic method
allowed for calling findBy[Property](), findOneBy[Property]()
and countBy[Property](). Those methods are now implemented
with a slightly different method signature but with type
declaration and phpstan support.

The naming follows the one of doctrine/orm which is
implemented exactly the same.

The new methods are:

- findBy(array $criteria, ...): QueryResultInterface
- findOneBy(array $criteria, ...): object|null
- count(array $criteria, ...): int

Using the magic methods is deprecated but will only
trigger an E_USER_DEPRECATED error as of version 13.

Releases: main
Resolves: #100071
Change-Id: I1b51dd5b583a225a8a4d6f948810b1e5ef5d2257
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78014
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>