[TASK] Remove Extbase Bootstrap::persistenceClasses
Property public static array $persistenceClasses
has been introduced with #87623 in v10 and never
been used. It seems this was an accidential left
over from a previous version of the patch, or
something like that.
We remove that unused property without further
deprecation since there was never a reason to
use it.

Change-Id: I4b08186e3c190ef564ce397739a7a53a53727834
Resolves: #100075
Related: #87623
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78018
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Torben Hansen <derhansen@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>