[BUGFIX] Respect TCA field foreign_default_sortby by extbase
Inside the extbase context child relations are always ordered
by the foreign_sortby field if set or by order of creation.
Even if the field foreign_default_sortby is set, it only
impacts the order of the backend view.

This patch brings the order of extbase results inline with
the order of the backend view.

Resolves: #64197
Releases: main, 11.5
Change-Id: I582576b5ac3741e2bbc4107664140fd9a2f63a16
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77985
Reviewed-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Alexander Schnitzler <git@alexanderschnitzler.de>