[BUGFIX] Fix type error in UnableToLinkException->getLinkText
The type hint indicates, that $linkText should be of type string.
However this is not always the case. In TYPO3 v11, it can contain
an instance of LinkResult.
The code has been refactored in TYPO3 v12 and since the
potential code has been removed from ContentObjectRenderer this
fix is to be considered temporary.

Resolves: #100106
Releases: 11.5
Change-Id: I61ba89651c77496ec5cc986a435a758466b03067
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78074
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>