[BUGFIX] Ensure `uid` is not updated when updating resources
When updating entries, the `updateFields` will provide `uid`, which
works fine in MySQL but might lead to problems when working with
stricter DBMS.

Such an example of a DBMS not liking the update of an identity field,
is SQL Server.

Now, I know that SQL Server will not be supported in TYPO3 v12 going
onward, but on the off-chance that somebody has another DBMS than MySQL
or has DBMS that will break when trying to update an identify field,
this bugfix might still be valuable.

The bug exists back to TYPO3 10 at least.

Releases: main, 12.4, 11.5
Resolves: #100175
Change-Id: Id4dd90579191da0cb770490256087209e749deb4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80067
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>