[FEATURE] Allow overriding button text for TCA type=file
With TCA, it is possible for type=file to change
the label to a different label depending on the field.

This is useful if you want to change the label
to "Choose poster image", if e.g. the selection
of file extensions is changed.

This is done via
* [config][appearance][createNewRelationLinkTitle]

Now, the other two buttons can also be changed:
* [config][appearance][uploadFilesLinkTitle] (defaults to "Select & upload files)
* [config][appearance][addMediaLinkTitle] (defaults to "Add media by URL")

Resolves: #65020
Releases: main
Change-Id: Icc2fee1c1c857aac667130a16bb8d5195dcb50e6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78190
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>