[TASK] Separate concerns in EXT:scheduler
This change finally cleans up logic within
the scheduler extension, namely to separate
concerns regarding reading from and writing to
the database.

All of these methods are moved to a Repository
class in order to handle database access into
one class, and to untangle the tight coupling
between the Scheduler class and the Task class.

In addition, the AbstractTask will become
a lot smaller, as it only deals with the actual
execution of the code in the future.

A new TaskValidator is added to validate if a
task can be executed.

Resolves: #100247
Releases: main
Change-Id: I6e74956bdffaf5c3197476456d8ba01dd0b2a97f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78197
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>