[TASK] Update lit JavaScript dependencies
In addition to the update we are dropping the version number
in the directory structure. This behavior was unique to the lit
dependencies, but the core itself does not provide a mapping
to require a specific version. The grunt copy task was adjusted
we are now also updating the resources when new versions
are required and we are cleaning the directories before
adding new files.

Executed commands:
npm i --save-dev \
	grunt-contrib-clean@^2.0.1
npm i --save \
	@lit/reactive-element@^1.6.1 \
	lit@^2.7.0 \
	lit-element@^3.3.0 \
	lit-html@^2.7.0
npm run build

Resolves: #100245
Releases: main
Change-Id: I5833a3f314d54b4bdc2a4cec8936954d7710ba38
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78208
Tested-by: core-ci <typo3@b13.com>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>