[TASK] Get rid of ColumnMap::$propertyName
The ColumnMap only carried the propertyName to be
queried by the DataMap (of a class) to have a key
for its internal stack of ColumnMaps.

The overall goal is to separate core (TCA columns)
logic from Extbase (entity properties) logic. A
ColumnMap should only describe the actual state of
the TCA, prepared for Extbase to read but without
enriched data from entities.

Releases: main
Resolves: #100274
Change-Id: Ife29d8dc89bdca07cb334095bd4048aaf8d6af9e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78233
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>