[BUGFIX] Respect default value for confirmation in ext:install
The default value for a confirmation is currently not respected
in the install tool on UI level and a confirmation dialog will
always show the "deny" button as selected. On CLI level the
default value is however taken into account.

This change sets the default value of the confirm dialog buttons
depending on the `defaultValue` property of the confirmation. If
the `defaultValue` is `true`, the "confirm" button is selected,
otherwise the "deny" button is selected.

Resolves: #100291
Releases: main, 11.5
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Iee6e89c61cc5ad7117f7e565b42d39bd82ce26f9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78248
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>