[TASK] Provide PSR-14 `RedirectWasHitEvent` as webhook message
With #99629 outgoing webhooks has been implemented. This also
provide a simple way to create custom webhook messages for
any PSR-14 event. Therefore, the `RedirectWasHit` event can
be used to provide an additional webhook type and showcase
the needed steps to add this for a core PSR-14 event.

This change provides a webhook message for the already
existing PSR-14 `RedirectWasHitEvent` of `ext:redirects`.

Resolves: #100305
Related: #99629
Releases: main
Change-Id: If63a665571402f9491c8a576516ec25080e7c31c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78261
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>