[TASK] Do not accept empty password in backend password reset
If the global password policy has been disabled and the backend
password function is used, it is possible to submit an empty
password, if the user manually removes the `required` attribute
from the password fields in the password reset form.

This change ensures, that an empty password will not be accepted
by the backend password reset function.

Resolves: #100295
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I20f6fa7f81650964e38e9b2bbb436cf3e06dbf16
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78324
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>