[BUGFIX] Check if requested start module is registered before resolution
The change in #100342 was not complete and might lead to exceptions
when a regular route name not matching a module is passed as the
start module, e.g. when reloading the browser while
EditDocumentController is open.

This is now guarded by a check whether whatever is coming in is an
actual registered module before attempting to resolve potential module
aliases.

Resolves: #100352
Related: #100342
Releases: main
Change-Id: I71805b147df3eb9421d0fce36940dd3873edef4c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78326
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>