[BUGFIX] Support @import 'EXT:.../TypoScript/*.setup.typoscript'
While a pattern like "@import 'EXT:.../TypoScript/*.typoscript'"
works already, "@import 'EXT:.../TypoScript/*.setup.typoscript'"
did not.

The patch adapts the wildcard resolver to deal with this, the
code is hopefully easier to follow along the way. A functional
test is added to cover above scenario.

The "TypoScript syntax changes" is tuned to clarify which
syntax for @import is supported, including a list of valid
and invalid examples.

Resolves: #100205
Related: #97816
Related: #90146
Related: #82812
Releases: main
Change-Id: Ie36e25971788790f0a89a3e368e4258d4e11140c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78331
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>