[TASK] Clean up avatar class
In a list of backend users, using the avatar viewhelper, the
instantiated Avatar class would always be the same instance.
Each time, getImage() is called, the avatar providers would be
initiated again, flooding the $avatarProviders variable with
superfluous providers. - Which meant, when rendering a list of
just 50 backend users, we suddenly have 1275 calls to the avatar
provider class, instead of just the expected 50.

Resolves: #100366
Releases: main
Change-Id: I8b2eed49b47853c52a4350eb06e671081e90e373
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78339
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>