[TASK] Avoid withConsecutive() in RequestHandlerTest
Even though the RequestHandlerTest unit tests
are an aweful mocking party, we can't simply
throw them away since the class is so important.

They should still be turned into functional
tests later, to be more useful, maybe together
with a refactoring of that class.

The patch just changes the tests to avoid
withConsecutive(), which does not improve
the situation, but a bigger refactoring
is currently a bit too much with our current
goal of "just" adding phpunit 10 compatibility.

Resolves: #100410
Related: #100249
Releases: main
Change-Id: I8b09013b826ed56e0c6e96eade4c9f4a61346273
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78379
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>