[BUGFIX] Add external media CSP declarations to frontend scope
Previously external media CSP declarations (e.g. to embed YouTube
or Vimeo assets in an iframe) was only applied to the backend scope,
when `ext:filelist` was installed. However, the same declarations
are required in the frontend scope for actually rendering those
external media content elements.

The initiating components `YouTubeRenderer` and `VimeoRenderer` are
located in `ext:core`. That's why the CSP declarations are moved
there as well.

Resolves: #100446
Releases: main
Change-Id: I4bb11f106539cbbc536e9a0ef7f77eb4e727d753
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78459
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>