[TASK] Remove BackendModuleControllerTest
This test is tricky since the subject triggers
session_save_path() which is limited to one
call per php process and emits a warning when
called multiple times.

This was mitigated by adding an own exception
handler in the test setup, but this solution
fails with phpunit 10, which is more picky
and can't be tricked easily.

The value added by the test is not *that*
high and the overall setup is complex, so
in this case, it's better to remove the test.

Resolves: #100472
Related: #100249
Releases: main
Change-Id: Iea0d4128cce9d804ce3d634e5095b19f5c9d8bed
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78480
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>