[TASK] Slightly improve belog rendering
We are introducing new table columns for avatar, time and datetime
in preparation for more adoptions. The avatar column also removes the
space on the right to bring it closer to the name of the user while
having a fixed size for medium avatars.

To avoid column jumping we also limit the size of the icon column to
the width of a small icon that is normally used in the tables in the
backend. The new time and datetime column classes are utilizing the
character unit and ensuring that the value is not wrapped to ensure a
better reading experience across multiple tables.

The button sizes in the table are now aligned with the list module and
the delete actions now feature an icon in addition to the text.

- Unnecessary/meaningless table headers for actions and icons are
  removed
- Unnecessary div blocks for the grouped log entries are removed.
- Unnecessary invisible divider was removed

Resolves: #100511
Releases: main
Change-Id: I15b4e42d31fc9426cfc20de36b39478eabb0ba33
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78519
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>