[BUGFIX] Move flex form inline children along with its parent
When a record is moved to a different page, flex form inline
children have to be moved to that pid as well, just like "casual"
inline children. Current DataHandler does not do this. This can
lead to falsely deleted records if an old source page is later
deleted by an editor.

The patch establishes fixes and test scenarios for flexform inline
to see if children are properly moved when a parent is moved. The
change goes along with an additional refactoring within moveRecord(),
aiming to make everything more obvious and less convoluted. The test
.csv files have some `@todo` comments that outline fishy sorting and
workspace related details. Those should be elaborated on with
upcoming changes.

Resolves: #106407
Related: #100518
Releases: main, 13.4
Change-Id: If7d164ff0438ce5df4903c7dc964cdf7af3caa68
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88680
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>